feat: Introduce modular chat hooks for improved structure and maintainability, including connection, messages, conversations, groups, and read status management
